pyproject.toml
# Build version: 1  # Increment this to force a fresh build
[tool.poetry]
name = "djangify-backend"
version = "0.1.0"
description = "Django Portfolio Backend"
authors = ["Diane Corriette"]
readme = "README.md"
packages = [{include = "config"}, {include = "apps"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
django-summernote = "^0.8.20.0"
whitenoise = "^6.8.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
dj-database-url = "^2.1.0"
pillow = "^11.0.0"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
python-decouple = "^3.8"
python-environ = "^0.4.54"
cloudinary = "^1.41.0"
django-cloudinary-storage = "^0.3.0"

[build-system]
requires = [
    "poetry-core>=1.0.0",
    "setuptools>=45",
    "wheel"
]
build-backend = "poetry.core.masonry.api"
